refactor(inferno,galaxy): Share code for ISO drivers
I guess this also a feat for `galaxy`: Add (untested) support for galaxy to load-exec CSOv2, JSO, ZSO and DAX iso formats
