+++ cef/galaxy/galaxy.h
@@ -24,55 +24,12 @@
 #include <pspsdk.h>
 #include <stdio.h>
 
-// Read Argument for Kernel Extend Stack
-typedef struct IoReadArg {
-	u32 offset;
-	u8 *address;
-	u32 size;
-} IoReadArg;
-
-// CSO Header
-typedef struct CISO_header {
-	unsigned char magic[4];
-	unsigned int header_size;
-	uint64_t total_bytes;
-	unsigned int block_size;
-	unsigned char ver;
-	unsigned char align;
-	unsigned char rsv_06[2];
-} CISO_header;
-
 #define NP9660_DATA_1 (0x00005BB4 - 0x00005BA4 + 0x00000188 + 0x00008900)
 #define NP9660_DATA_2 (0x00005BBC - 0x00005BA4 + 0x00000188 + 0x00008900)
 #define NP9660_DATA_3 (0x00005BD0 - 0x00005BA4 + 0x00000188 + 0x00008900)
 #define NP9660_DATA_4 (0x00005BD8 - 0x00005BA4 + 0x00000188 + 0x00008900)
 #define NP9660_DATA_5 (0x00000114 + 0x00008900)
 #define NP9660_ISO_FD (0x00000188 + 0x00008900)
-
-// ISO Sector Size
-#define ISO_SECTOR_SIZE 0x800
-
-// CSO Buffer Size
-#define CISO_IDX_BUFFER_SIZE 0x200
-#define CISO_DEC_BUFFER_SIZE 0x2000
-
-// Get Disc Sector Count
-int get_total_block(void);
-
-// Read Raw ISO Data
-int read_raw_data(unsigned char * addr, unsigned int size, unsigned int offset);
-
-// Read CSO Disc Sector
-int read_cso_sector(unsigned char * addr, int sector);
-
-// Custom CSO Sector Reader
-int read_cso_data(unsigned char * addr, unsigned int size, unsigned int offset);
-
-// Custom ISO Sector Reader Wrapper
-int iso_read(struct IoReadArg * args);
-
-// Open CSO File (subcall from open_iso)
-int cso_open(SceUID fd);
 
 // Open ISO File
 int open_iso(void);
