+++ cef/inferno/inferno.h
@@ -18,9 +18,7 @@
 #ifndef INFERNO_H
 #define INFERNO_H
 
-#define ISO_SECTOR_SIZE 2048
-#define CISO_IDX_BUFFER_SIZE 0x200
-#define CISO_DEC_BUFFER_SIZE 0x2000
+#include <iso_common.h>
 
 #define MAX_FILES_NR 8
 
@@ -32,12 +30,6 @@
 		} \
 	} while ( 0 )
 
-typedef struct IoReadArg {
-	u32 offset; // 0
-	u8 *address; // 4
-	u32 size; // 8
-} IoReadArg;
-
 extern u32 psp_model;
 extern u32 psp_fw_version;
 extern PspIoDrv g_iodrv;
@@ -46,15 +38,10 @@ extern SceUID g_umd_cbid;
 extern int g_umd_error_status;
 extern int g_drive_status;
 
-extern char g_iso_fn[255];
 extern int g_game_fix_type;
 extern SceUID g_drive_status_evf;
 extern void *g_sector_buf;
 extern SceUID g_umd9660_sema_id;
-extern int g_iso_opened;
-extern SceUID g_iso_fd;
-extern int g_total_sectors;
-extern struct IoReadArg g_read_arg;
 extern int g_disc_type;
 extern unsigned char umd_seek;
 extern unsigned char umd_speed;
@@ -65,8 +52,6 @@ extern void sceUmdSetDriveStatus(int status);
 
 extern int power_event_handler(int ev_id, char *ev_name, void *param, int *result);
 
-extern int iso_open(void);
-extern int iso_read(struct IoReadArg *args);
 extern int iso_cache_read(struct IoReadArg *args);
 extern int iso_read_with_stack(u32 offset, void *ptr, u32 data_len);
