+++ cef/inferno/main.c
@@ -40,9 +40,6 @@ u32 psp_fw_version;
 
 int sceKernelSetQTGP3(void *data);
 
-// 00002790
-char g_iso_fn[255] = {0};
-
 // 0x00002248
 u8 g_umddata[16] = {
 	0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
@@ -61,9 +58,9 @@ PspSysEventHandler g_power_event = {
 // 00000090
 int setup_umd_device(void) {
 	memset(g_iso_fn, 0, sizeof(g_iso_fn));
-    strncpy(g_iso_fn, GetUmdFile(), sizeof(g_iso_fn));
+	strncpy(g_iso_fn, GetUmdFile(), sizeof(g_iso_fn));
 
-    infernoSetDiscType(sctrlSEGetDiscType());
+	infernoSetDiscType(sctrlSEGetDiscType());
 
 	int ret = sceIoAddDrv(&g_iodrv);
 
@@ -98,6 +95,8 @@ int module_start(SceSize args, void* argp) {
 	if (ret < 0) {
 		return ret;
 	}
+
+	logmsg3("[INFO]: UMD File: %s\n", g_iso_fn);
 
 	ret = init_inferno();
